net: Move aiohttp session management to http
Relates to #1078
