+++ docs/api/pyatv.html
@@ -60,7 +60,7 @@ from pyatv.dmap.pairing import DmapPairingHandler
 from pyatv.mrp import setup as mrp_setup
 from pyatv.mrp.pairing import MrpPairingHandler
 from pyatv.raop import setup as raop_setup
-from pyatv.support import net
+from pyatv.support import http
 from pyatv.support.facade import FacadeAppleTV, SetupMethod
 from pyatv.support.scan import BaseScanner, MulticastMdnsScanner, UnicastMdnsScanner
 
@@ -119,7 +119,7 @@ async def connect(
     if config.identifier is None:
         raise exceptions.DeviceIdMissingError(&#34;no device identifier&#34;)
 
-    session_manager = await net.create_session(session)
+    session_manager = await http.create_session(session)
     atv = FacadeAppleTV(config, session_manager)
 
     for service in config.services:
@@ -163,7 +163,7 @@ async def pair(
     if handler is None:
         raise RuntimeError(f&#34;missing implementation for {protocol}&#34;)
 
-    return handler(config, await net.create_session(session), loop, **kwargs)</code></pre>
+    return handler(config, await http.create_session(session), loop, **kwargs)</code></pre>
 </details>
 </section>
 <section>
@@ -222,7 +222,7 @@ async def pair(
     if config.identifier is None:
         raise exceptions.DeviceIdMissingError(&#34;no device identifier&#34;)
 
-    session_manager = await net.create_session(session)
+    session_manager = await http.create_session(session)
     atv = FacadeAppleTV(config, session_manager)
 
     for service in config.services:
@@ -275,7 +275,7 @@ async def pair(
     if handler is None:
         raise RuntimeError(f&#34;missing implementation for {protocol}&#34;)
 
-    return handler(config, await net.create_session(session), loop, **kwargs)</code></pre>
+    return handler(config, await http.create_session(session), loop, **kwargs)</code></pre>
 </details>
 </dd>
 <dt id="pyatv.scan"><code class="name flex">
