+++ pyatv/dmap/__init__.py
@@ -35,7 +35,7 @@ from pyatv.interface import (
 )
 from pyatv.support import deprecated, net
 from pyatv.support.cache import Cache
-from pyatv.support.net import ClientSessionManager
+from pyatv.support.http import ClientSessionManager
 from pyatv.support.relayer import Relayer
 
 _LOGGER = logging.getLogger(__name__)
