+++ tests/conftest.py
@@ -4,7 +4,8 @@ from unittest.mock import Mock, patch
 import netifaces
 import pytest
 
-from pyatv.support.net import create_session, unused_port
+from pyatv.support.http import create_session
+from pyatv.support.net import unused_port
 
 from tests.fake_knock import create_knock_server
 from tests.utils import stub_sleep, unstub_sleep
